+++ core/src/main/java/tech/pegasys/web3signer/core/Runner.java
@@ -15,6 +15,7 @@ package tech.pegasys.web3signer.core;
 import static tech.pegasys.web3signer.core.service.http.OpenApiOperationsId.UPCHECK;
 
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.web3signer.common.ApplicationInfo;
 import tech.pegasys.web3signer.core.config.ClientAuthConstraints;
 import tech.pegasys.web3signer.core.config.Config;
 import tech.pegasys.web3signer.core.config.TlsOptions;
@@ -64,6 +65,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.core.config.Configurator;
 import org.apache.tuweni.net.tls.VertxTrustOptions;
+import org.hyperledger.besu.metrics.StandardMetricCategory;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
 
 public abstract class Runner implements Runnable {
@@ -99,6 +101,7 @@ public abstract class Runner implements Runnable {
         createArtifactSignerProvider(vertx, metricsSystem);
 
     try {
+      createVersionMetric(metricsSystem);
       metricsEndpoint.start(vertx);
 
       try {
@@ -170,6 +173,13 @@ public abstract class Runner implements Runnable {
       metricsEndpoint.stop();
       LOG.error("Failed to initialise application", e);
     }
+  }
+
+  private void createVersionMetric(MetricsSystem metricsSystem) {
+    metricsSystem
+        .createLabelledGauge(
+            StandardMetricCategory.PROCESS, "release", "Release information", "version")
+        .labels(() -> 1, ApplicationInfo.version());
   }
 
   private VertxOptions createVertxOptions(final MetricsSystem metricsSystem) {